🧹 Refactor EncodedValueTransformer logic into EncodedValueWriter
Extracted the traversal and writing logic for DEX encoded values from the `EncodedValueTransformer` inner class in `IndexMap.java` into a new abstract helper class `EncodedValueWriter` in the `mod.agus.jcoderz.dex` package.

What:
- Created `EncodedValueWriter.java` as an abstract base class for writing encoded values.
- Moved the `transform`, `transformAnnotation`, and `transformArray` methods from `IndexMap.java` to `EncodedValueWriter.java`.
- Defined hook methods (`adjustString`, `adjustType`, etc.) in `EncodedValueWriter` to allow customization of index mapping.
- Updated `IndexMap.EncodedValueTransformer` to extend `EncodedValueWriter` and override the adjustment hooks.

Why:
- Improves code maintainability and reusability by separating generic DEX value writing logic from project-specific index mapping logic.
- Reduces the size and complexity of `IndexMap.java`.

Verification:
- Manually verified that the extracted logic is identical to the original implementation.
- Successfully compiled the modified and new files using `javac` with the appropriate classpath and sourcepath.

Result:
- Cleaner `IndexMap.java` and a reusable `EncodedValueWriter` class for future DEX manipulation tasks.
